pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "flora"
version = "0.1.0"
description = "Autonomous herb garden agent powered by Claude"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "aiosqlite>=0.20.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "jinja2>=3.1.4",
    "python-telegram-bot>=21.0",
    "apscheduler>=3.10.4",
    "gpiozero>=2.0",
    "miflora>=0.7.2",
    "bleak>=0.22.0",
    "python-kasa>=0.7.4",
    "adafruit-circuitpython-sht31d>=2.3.27",
    "adafruit-circuitpython-bh1750>=1.1.4",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24.0", "mypy>=1.11.0"]

[project.scripts]
flora = "flora.main:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
